keyper/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-data-tables
azure-functions
azure-identity
azure-storage-blob
azure-storage-queue
azure-ai-vision-imageanalysis